requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
django-cors-headers==3.2.0
django-money==1.0
django==3.0
djangorestframework==3.10.3
drf-nested-routers==0.91
py-moneyed==0.8.0
pytz==2019.3
sqlparse==0.3.0
